% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_exp_time.R
\name{add_exp_time}
\alias{add_exp_time}
\title{Add experimental time information}
\usage{
add_exp_time(dataset, exp_hour_var = "exp.hour")
}
\arguments{
\item{dataset}{dataframe}

\item{exp_hour_var}{exp.hour}
}
\value{
dataframe with additional columns for exp.day and exp.min
}
\description{
Use this function to add experiment and minute number based on your newly calculated hours
}
\examples{
\dontrun{
add_exp_time(df1)
add_exp_time(df1, exp_hour_var = "exp.hour")
}


}
